Instruction Inputs are now Variables.
To distinguish between a literal and and reference all the inputs and
outputs across the system needed to change to Variables.

This effected across the majority of the system, but has successfully
been tested on MPS code.

Instruction IDs have been changed to long to capture the data from MPS.

+++ src/main/java/executable/instructions/Detect.java
@@ -7,11 +7,11 @@ import substance.Substance;
  */
 public class Detect extends Instruction {
 
-	public Detect(int id, String name) {
+	public Detect(long id, String name) {
 		super(id, name, Detect.class);
 	}
 
-	public Detect(int id) {
+	public Detect(long id) {
 		super(id, Detect.class);
 	}
